Move directory base objects
