+++ src/telliot_core/directory/base.py
@@ -1,16 +1,49 @@
+from dataclasses import dataclass
+from dataclasses import field
 from typing import Any
-from typing import Dict
-from typing import TypedDict
+from typing import List
+from typing import Optional
 
 
-# Container for contract access information
-class ContractInfo(TypedDict):
+@dataclass
+class ContractInfo:
+    org: str
+    name: str
+    chain_id: int
     address: str
-    abi: Dict[Any, Any]
+    abi: Optional[List[Any]] = field(repr=False)
 
 
-# Mapping of contract name to Contract Info
-ContractSet = Dict[str, ContractInfo]
+@dataclass
+class ContractDirectory:
+    # Private directory storage
+    _contents: List[ContractInfo] = field(default_factory=list)
 
-# Mapping of chain_id to set of contracts deployed on that chain
-Directory = Dict[int, ContractSet]
+    def add_contract(self, info: ContractInfo) -> None:
+        self._contents.append(info)
+
+    def find(
+        self,
+        *,
+        org: str = "tellor",
+        name: Optional[str] = None,
+        address: Optional[str] = None,
+        chain_id: Optional[int] = None,
+    ) -> List[ContractInfo]:
+        result = []
+        for info in self._contents:
+            if info.org != org:
+                continue
+            if chain_id is not None:
+                if chain_id != info.chain_id:
+                    continue
+            if name is not None:
+                if info.name is not name:
+                    continue
+            if address is not None:
+                if info.address is not address:
+                    continue
+
+            result.append(info)
+
+        return result
